chore(bot:types): convert types to declaration files & move outside of src
