+++ apps/bot/src/events/ready.ts
@@ -4,9 +4,9 @@ import { dirname, join } from "node:path";
 import { fileURLToPath } from 'node:url';
 import { importDir } from "utils";
 import { cache } from "..";
+import { Command } from "../../types/Command";
 import { updateCachedReactionNotifiers } from "../handlers/reactionNotifications";
 import { BOT_PORT, server } from "../server";
-import { Command } from "../types/Command";
 import { makeEvent } from "../utils/events/makeEvent";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
