+++ apps/bot/src/utils/events/makeEvent.ts
@@ -1,5 +1,6 @@
 import { ClientEvents } from "discord.js";
-import { DiscordEvent } from "../../types/DiscordEvent";
+import { DiscordEvent } from "../../../types/DiscordEvent";
+
 
 export const makeEvent = <K extends keyof ClientEvents>(event: DiscordEvent<K>) => {
     return event;
