+++ apps/bot/types/Command.d.ts
