+++ apps/bot/types/DiscordEvent.d.ts
