+++ apps/bot/types/GuildSettings.d.ts
